set up sass, initial css files split
