+++ package.json
@@ -19,6 +19,7 @@
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "react-hot-toast": "^2.2.0",
+    "sass": "^1.52.1",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
