+++ styles/Form.module.scss
@@ -1,0 +1,66 @@
+
+.formBlock {
+    margin-bottom: 1rem;
+  }
+  
+  .label {
+    display: block;
+    margin-bottom: 0.2rem;
+  }
+  
+  .input{
+    width: 100%;
+    padding: 1rem;
+    background: var(--color-pink-light);
+    border: 1px solid black;
+    border-radius: 5px;
+    font-family: inherit;
+    font-size: var(--font-size-normal);
+  }
+    .input:disabled{
+      background: var(--color-gray-light);
+      cursor: not-allowed;
+    }
+  
+  .button{
+    width: 100%;
+    padding: 1rem;
+    color: var(--color-gray-dark);
+    font-size: var(--font-size-normal);
+    font-weight: bold;
+    cursor: pointer;
+  }
+    .button:disabled {
+      color: var(--color-gray-light);
+      cursor: not-allowed;
+    }
+  .buttonGreen {
+    background: var(--color-green);
+    border: 1px solid var(--color-green-dark);
+  }
+    .buttonGreen:active {
+      background: var(--color-green-dark);
+    }
+    .buttonGreen:disabled {
+      background: var(--color-green-light);
+    }
+  .buttonBlue {
+    background: var(--color-blue);
+    border: 1px solid var(--color-blue-dark);
+  }
+    .buttonBlue:active {
+      background: var(--color-blue-dark);
+    }
+    .buttonBlue:disabled {
+      background: var(--color-blue-light);
+    }
+  .buttonRed {
+    background: var(--color-red);
+    border: 1px solid var(--color-red-dark);
+  }
+    .buttonRed:active {
+      background: var(--color-red-dark);
+    }
+    .buttonRed:disabled {
+      background: var(--color-red-light);
+    }
