+++ styles/Groups.module.scss
@@ -1,0 +1,12 @@
+.groupList {
+    margin: 1rem 0;
+  }  
+.groupItem {
+  margin-bottom: 1rem;
+}
+.groupMembers {
+  font-style: italic;
+}
+.groupDetails {
+  margin-bottom: 1rem;
+}
