--- styles/Home.module.css
@@ -1,237 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh; /* TODO: ensure the min height is OK */
-  width: min(100%, 800px);
-  margin: auto;
-  background: var(--color-gray-dark);
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  padding: 0.5rem 1rem;
-  border-bottom: 2px solid var(--color-pink-light);
-}
-.title {
-  flex: 1;
-  margin: 0;
-  font-size: var(--font-size-biggest);
-  color: var(--color-primary);
-}
-.title a {
-  text-decoration: none;
-}
-
-.pageTitle {
-  margin-bottom: 1rem;
-  font-size: var(--font-size-big);
-}
-
-.main {
-  width: 100%;
-  flex-grow: 1;
-  padding: 1rem 2rem;
-}
-
-.footer {
-  display: flex;
-  width: 100%;
-  padding: 0.5rem 1rem;
-  border-top: 2px solid var(--color-pink-light);
-}
-.nav {
-  margin-left: auto;
-  display: flex;
-  justify-content: flex-end;
-  align-content: center;
-  gap: 0.5rem;
-}
-
-.authGuardianContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-  .authGuardianContainer h2 {
-    font-size: var(--font-size-biggest);
-    margin-bottom: 1rem;
-  }
-  .authGuardianContainer h3 {
-    margin-bottom: 1rem;
-    font-size: var(--font-size-normal);
-  }
-
-.formBlock {
-  margin-bottom: 1rem;
-}
-
-.label {
-  display: block;
-  margin-bottom: 0.2rem;
-}
-
-.input{
-  width: 100%;
-  padding: 1rem;
-  background: var(--color-pink-light);
-  border: 1px solid black;
-  border-radius: 5px;
-  font-family: inherit;
-  font-size: var(--font-size-normal);
-}
-  .input:disabled{
-    background: var(--color-gray-light);
-    cursor: not-allowed;
-  }
-
-.button{
-  width: 100%;
-  padding: 1rem;
-  color: var(--color-gray-dark);
-  font-size: var(--font-size-normal);
-  font-weight: bold;
-  cursor: pointer;
-}
-  .button:disabled {
-    color: var(--color-gray-light);
-    cursor: not-allowed;
-  }
-.buttonGreen {
-  background: var(--color-green);
-  border: 1px solid var(--color-green-dark);
-}
-  .buttonGreen:active {
-    background: var(--color-green-dark);
-  }
-  .buttonGreen:disabled {
-    background: var(--color-green-light);
-  }
-.buttonBlue {
-  background: var(--color-blue);
-  border: 1px solid var(--color-blue-dark);
-}
-  .buttonBlue:active {
-    background: var(--color-blue-dark);
-  }
-  .buttonBlue:disabled {
-    background: var(--color-blue-light);
-  }
-.buttonRed {
-  background: var(--color-red);
-  border: 1px solid var(--color-red-dark);
-}
-  .buttonRed:active {
-    background: var(--color-red-dark);
-  }
-  .buttonRed:disabled {
-    background: var(--color-red-light);
-  }
-
-.tasksList {}
-.tasksListItem {
-  display: grid;
-  grid-template-columns: 4fr 1fr;
-  margin-bottom: 1rem;
-  width: 100%;
-  padding: 1rem;
-  border: 1px solid var(--color-gray);
-  background: var(--color-gray);
-  cursor: pointer;
-}
-  .tasksListItem:hover{
-    border: 1px solid var(--color-pink-light);
-  }
-.tasksListName {
-  margin-bottom: 1rem;
-}
-.tasksListDueDate {
-  font-style: italic;
-  font-size: var(--font-size-small);
-}
-.tasksListConfirmation {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.taskDetails {
-  margin-bottom: 1rem;
-}
-.taskActions {
-  padding: 0 2rem;
-}
-  .taskActions button {
-    margin-bottom: 1rem;
-  }
-
-.groupList {
-    margin: 1rem 0;
-  }  
-.groupItem {
-  margin-bottom: 1rem;
-}
-.groupMembers {
-  font-style: italic;
-}
-.groupDetails {
-  margin-bottom: 1rem;
-}
-
-.tabsContainerGroup {
-  display: flex;
-}
-.tabsContainer {
-  display: flex;
-  margin-bottom: 1rem;
-}
-  .tabsContainer:first-child {
-    flex: 1;
-  }
-.tab {
-  padding: 0 0.5rem;
-  border: none;
-  background: none;
-  font-size: var(--font-size-normal);
-  color: white;
-  cursor: pointer;
-}
-  .tab:not(:last-child) {
-    border-right: 1px solid white;
-  }
-  .tabCurrent {
-    text-decoration: underline;
-  }
-
-.spinner {
-  animation: rotate 2s linear infinite;
-  width: 30px;
-  height: 30px;
-}
-.spinnerPath {
-  stroke-linecap: round;
-  animation: dash 1.5s ease-in-out infinite;
-}
-@keyframes rotate {
-  100% {
-    transform: rotate(360deg);
-  }
-}
-@keyframes dash {
-  0% {
-    stroke-dasharray: 1, 150;
-    stroke-dashoffset: 0;
-  }
-  50% {
-    stroke-dasharray: 90, 150;
-    stroke-dashoffset: -35;
-  }
-  100% {
-    stroke-dasharray: 90, 150;
-    stroke-dashoffset: -124;
-  }
-}
