+++ styles/Layout.module.scss
@@ -1,0 +1,67 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh; /* TODO: ensure the min height is OK */
+    width: min(100%, 800px);
+    margin: auto;
+    background: var(--color-gray-dark);
+  }
+  
+  .header {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    padding: 0.5rem 1rem;
+    border-bottom: 2px solid var(--color-pink-light);
+  }
+  .title {
+    flex: 1;
+    margin: 0;
+    font-size: var(--font-size-biggest);
+    color: var(--color-primary);
+  }
+  .title a {
+    text-decoration: none;
+  }
+  
+  .pageTitle {
+    margin-bottom: 1rem;
+    font-size: var(--font-size-big);
+  }
+  
+  .main {
+    width: 100%;
+    flex-grow: 1;
+    padding: 1rem 2rem;
+  }
+  
+  .footer {
+    display: flex;
+    width: 100%;
+    padding: 0.5rem 1rem;
+    border-top: 2px solid var(--color-pink-light);
+  }
+  .nav {
+    margin-left: auto;
+    display: flex;
+    justify-content: flex-end;
+    align-content: center;
+    gap: 0.5rem;
+  }
+
+  .authGuardianContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+    .authGuardianContainer h2 {
+      font-size: var(--font-size-biggest);
+      margin-bottom: 1rem;
+    }
+    .authGuardianContainer h3 {
+      margin-bottom: 1rem;
+      font-size: var(--font-size-normal);
+    }
