+++ styles/Tasks.module.scss
@@ -1,0 +1,36 @@
+.tasksListItem {
+    display: grid;
+    grid-template-columns: 4fr 1fr;
+    margin-bottom: 1rem;
+    width: 100%;
+    padding: 1rem;
+    border: 1px solid var(--color-gray);
+    background: var(--color-gray);
+    cursor: pointer;
+  }
+    .tasksListItem:hover{
+      border: 1px solid var(--color-pink-light);
+    }
+  .tasksListName {
+    margin-bottom: 1rem;
+  }
+  .tasksListDueDate {
+    font-style: italic;
+    font-size: var(--font-size-small);
+  }
+  .tasksListConfirmation {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .taskDetails {
+    margin-bottom: 1rem;
+  }
+  .taskActions {
+    padding: 0 2rem;
+  }
+    .taskActions button {
+      margin-bottom: 1rem;
+    }
+
